Revert "Changes to the compression script (#78)" (#80)
This reverts commit de12b740bbee9970b81ec347180afe992522af92.
