+++ src/cli.jl
@@ -1,8 +1,18 @@
-function common_args(s)
+using Stitch
+using ArgParse
+
+import Stitch as S
+import JSON
+
+
+function cli()
+    # parse argument --iterations
+    s = ArgParseSettings()
     @add_arg_table s begin
-        "--corpus"
-        help = "Corpus of programs"
-        arg_type = String
+        "--iterations"
+        help = "Number of iterations to run"
+        default = 1
+        arg_type = Int
     end
 
     @add_arg_table s begin
@@ -63,20 +73,24 @@ function common_args(s)
         help = "Disallow metavariables from being seqS"
         action = :store_true
     end
-end
 
-function gather_common_arguments(args)
+    args = parse_args(s)
+
     size_by_symbol_json = JSON.parse(args["size-by-symbol"])
 
-    corpus = Corpus([Program(parse(SExpr, p), i, i) for (i, p) in enumerate(JSON.parse(args["corpus"]))])
+    line = readline()
+    json = JSON.parse(line)
+    corpus = Corpus([Program(parse(SExpr, p), i, i) for (i, p) in enumerate(json)])
 
     size_by_symbol = Dict(Symbol(k) => Float32(v) for (k, v) in size_by_symbol_json)
     dfa_valid_root_states = Set([Symbol(s) for s in JSON.parse(args["dfa-valid-root-states"])])
-    kwargs = (;
+    abstractions, corpus, _ = compress(
+        corpus,
         dfa=load_dfa(args["dfa"]),
         autoexpand_head=true,
         verbose_best=true,
         allow_single_task=false,
+        iterations=args["iterations"],
         max_arity=args["max-arity"],
         match_sequences=true,
         size_by_symbol=size_by_symbol,
@@ -87,5 +101,22 @@ function gather_common_arguments(args)
         dfa_metavariable_allow_S=!args["dfa-metavariable-disallow-S"],
         dfa_metavariable_allow_seqS=!args["dfa-metavariable-disallow-seqS"],
     )
-    return corpus, kwargs
+    println("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
+    println(JSON.json([
+        Dict([
+            ("arity", abstraction.arity),
+            ("sym_arity", abstraction.sym_arity),
+            ("choice_arity", abstraction.choice_arity),
+            ("body", string(abstraction.body)),
+            ("dfa_root", abstraction.dfa_root),
+            ("dfa_metavars", abstraction.dfa_metavars),
+            ("dfa_symvars", abstraction.dfa_symvars),
+            ("dfa_choicevars", abstraction.dfa_choicevars),
+        ])
+        for abstraction in abstractions
+    ]))
+    println("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~")
+    println(JSON.json([string(p) for p in corpus.programs]))
 end
+
+cli()
